refactor(demos): 按功能模块拆分文件，日志降级为 Debug
- 拆分为 8 个文件：main / config / plugin / media + 4 个 handler 文件
  （handler_image / handler_video / handler_text / handler_search）
- slog.Info 统一降级为 slog.Debug（非必要插件）
- 移除 ensureDefaults，默认值由 main 提供（与 setu 一致）
